create ci package and move util there
